refactor(ai-engine): resolve #1731 — split query_expansion.py into query_expansion/ subpackage
Split the 32K query_expansion.py monolith into a focused subpackage
following the same pattern as agents/advanced_rag/ (#1709) and
agents/logic_auditor/ (#1708).

New structure:
  search/query_expansion/
  ├── __init__.py            # re-exports full public API (backward-compatible)
  ├── models.py              # ExpansionStrategy, ExpansionTerm, ExpandedQuery
  ├── domain_expander.py     # MinecraftDomainExpander
  ├── synonym_expander.py    # SynonymExpander
  ├── contextual_expander.py # ContextualExpander
  └── engine.py              # QueryExpansionEngine coordinator

Zero behavior changes — pure code reorganization. All 7 public symbols
(ExpansionStrategy, ExpansionTerm, ExpandedQuery, MinecraftDomainExpander,
SynonymExpander, ContextualExpander, QueryExpansionEngine) are re-exported
from __init__.py so existing 'from search.query_expansion import X' imports
continue to work unchanged.
